% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-string.R
\name{string.append}
\alias{string.append}
\title{Use Regex to Append Text}
\usage{
string.append(x, match, text, last = FALSE)
}
\arguments{
\item{x}{A character string. The text to search.}

\item{match}{A character string. The regex text to match.}

\item{text}{A character string. The text to append.}

\item{last}{A logical. Toggles whether to append only the
final match.}
}
\description{
\code{string.append} uses regular expressions to match a pattern,
 then appends \code{text} at the end of all matches.
}
