% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-string.R
\name{string.call}
\alias{string.call}
\title{Call String}
\usage{
string.call(x)
}
\arguments{
\item{x}{A character string. The text to search.}
}
\description{
\code{string.call} executes a string as an expression.
}
