% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-string.R
\name{string.collapse}
\alias{string.collapse}
\title{Collapse String}
\usage{
string.collapse(x)
}
\arguments{
\item{x}{A character string. The text to search.}
}
\description{
\code{string.collapse} converts a vector of character strings
 to a single character string.
}
