% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-string.R
\name{string.explode}
\alias{string.explode}
\title{Explode String}
\usage{
string.explode(x)
}
\arguments{
\item{x}{A character string. The text to search.}
}
\description{
\code{string.explode} converts one character string to a vector
 of individual character strings.
}
