% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-string.R
\name{string.extract}
\alias{string.extract}
\title{Use Regex to Extract Text}
\usage{
string.extract(x, match)
}
\arguments{
\item{x}{A character string. The text to search.}

\item{match}{A character string. The regex text to match.}
}
\description{
\code{string.extract} uses regular expressions to match a pattern,
 then returns all matches.
}
