% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-yuck.R
\name{yuck}
\alias{yuck}
\alias{yuck}
\alias{:=}
\title{Ye Upsold Comprehension Kit}
\usage{
":="(left, right)
}
\arguments{
\item{left}{The variable to which to assign the output.}

\item{right}{The for-loop(s) to parse.}
}
\description{
See Examples.
}
\examples{
\dontrun{
a := for(i in 1:5) i^2
a

a := for(i in 1:5) i \%in\% 1:4
a

a := for(i in 1:5) for(j in 1:7) i * j
matrix(a, 7, 5)
}
}
